OGC-12 : refactor Chat entity + add Message entity

+++ app/src/main/java/com/offgrid/coupler/data/entity/Chat.java
@@ -13,7 +13,7 @@ public class Chat {
     @PrimaryKey(autoGenerate = true)
     @NonNull
     @ColumnInfo(name = "id")
-    private Integer id;
+    private Long id;
 
     @NonNull
     @ColumnInfo(name = "title")
@@ -51,11 +51,11 @@ public class Chat {
 
 
     @NonNull
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(@NonNull Integer id) {
+    public void setId(@NonNull Long id) {
         this.id = id;
     }
